refactor: sendContactRequest is moved to mailing service, and reused in action
