+++ src/backend/mailing/mailing.service.ts
@@ -1,6 +1,36 @@
 import { mailersendClient, MailTemplateId } from '@/lib/mailersendClient'
 import { Recipient } from 'mailersend'
 
+export type ContactRequestEmailParams = {
+  senderEmail: string
+  senderFullName: string
+  recipientEmail: string
+  subject: string
+}
+
+export const sendContactRequest = async ({
+  senderEmail,
+  senderFullName,
+  subject,
+  recipientEmail,
+}: ContactRequestEmailParams) => {
+  try {
+    const config = {
+      fromEmail: senderEmail,
+      fromName: senderFullName,
+      subject: subject,
+    }
+    const recipients = [new Recipient(recipientEmail)]
+    await mailersendClient.sendMail({
+      recipients,
+      templateId: MailTemplateId.contactRequest,
+      config,
+    })
+  } catch (error) {
+    console.error('Error occured whilst sending contact request.', error)
+  }
+}
+
 export const sendProfileApprovedEmail = async (email: string) => {
   const templateId = MailTemplateId.profileApprovedNotification
