+++ src/components/ContactForm/ContactForm.tsx
@@ -1,4 +1,4 @@
-import { sendContactRequest } from '@/actions/mailing/sendContactRequest'
+import { contactRequestEmail } from '@/actions/mailing/contactRequestEmail'
 import { Button } from '@/components/Button/Button'
 import InputFormError from '@/components/InputFormError/InputFormError'
 import TextArea from '@/components/TextArea/TextArea'
@@ -25,7 +25,7 @@ export default function ContactForm({
       try {
         // Handle submit actions
         // console.log('Handle submit', values)
-        await sendContactRequest({
+        await contactRequestEmail({
           senderEmail: values.senderEmail,
           senderFullName: values.senderFullName,
           recipientEmail: userProfile.userEmail,
